backend/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
pip==24.2
pymongo==4.13.0
python-dotenv==1.1.0
Werkzeug==3.1.3